+++ source/RewardsService/rewardsservice/url_patterns.py
@@ -1,7 +1,13 @@
 from handlers.rewards_handler import RewardsHandler
-from handlers.user_handler import UserHandler
+from handlers.set_customer_handler import SetCustomerHandler
+from handlers.get_customer_handler import GetCustomer
+from handlers.delete_customer_handler import DeleteCustomer
+from handlers.get_all_customer_handler import AllCustomer
 
 url_patterns = [
     (r'/rewards', RewardsHandler),
-    (r'/user', UserHandler),
+    (r'/set', SetCustomerHandler),
+    (r'/get', GetCustomer),
+    (r'/del', DeleteCustomer),
+    (r'/all', AllCustomer),
 ]
